requirements.txt
pyodbc==4.0.39
pandas==2.1.0
openpyxl==3.1.2
python-dotenv==1.0.0
tqdm==4.66.1 